Silences warnings, fix bugs found that way (#499)
This patch silences:
* signed/unsigned comparisons (related to int / size_t)
* unused variables (some, there are more)
* uninitialized variable (implicitly mapped 'ptr'; added '= NULL')
* unused but set variable.
* 'missing braces around initializer' (gcc warned, g++ didn't)

Bug fixes:
* tests/5.0/simd/test_simd_nontemporal.c: Add missing 'omp' to pragma.
* tests/5.1/default/test_task_target_default_firstprivate.c: Likewise;
  don't map numerical constant 'N'.
* tests/5.0/target_update/test_target_update_mapper_from_discontiguous.c:
  Use 'i += 2' instead of no op '2+i' as loop step.

+++ tests/5.0/declare_mapper/test_declare_mapper_target_struct.c
@@ -28,7 +28,7 @@ typedef struct myvec{
 
 void init( myvec_t *s )
 { 
-  for(int i = 0; i < s->len; i++) 
+  for(size_t i = 0; i < s->len; i++)
     s->data[i] = i; 
 }
